+++ src/main/java/br/finax/services/InvoiceService.java
@@ -2,11 +2,14 @@ package br.finax.services;
 
 import br.finax.dto.InvoiceMonthValues;
 import br.finax.dto.InvoiceValues;
+import br.finax.exceptions.WithoutPermissionException;
 import br.finax.models.InvoicePayment;
 import br.finax.utils.FileUtils;
 import br.finax.utils.UtilsService;
+import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.Date;
@@ -25,6 +28,7 @@ public class InvoiceService {
     private final UtilsService utils;
     private final FileUtils fileUtils;
 
+    @Transactional(readOnly = true)
     public InvoiceMonthValues getInvoiceAndReleases(
             long creditCardId, String selectedMonth,
             Date firstDt, Date lastDt
@@ -38,6 +42,7 @@ public class InvoiceService {
         );
     }
 
+    @Transactional(readOnly = true)
     public InvoiceValues getValues() {
         return new InvoiceValues(
                 accountService.getBasicList(),
@@ -46,7 +51,10 @@ public class InvoiceService {
         );
     }
 
+    @Transactional
     public InvoicePayment savePayment(InvoicePayment payment) {
+        checkPermission(payment);
+
         if (payment.getId() != null) {
             final InvoicePayment invoicePayment = invoicePaymentService.findById(payment.getId());
 
@@ -57,12 +65,20 @@ public class InvoiceService {
         return invoicePaymentService.save(payment);
     }
 
+    @Transactional
     public void deletePayment(long invoicePaymentId) {
+        final InvoicePayment payment = invoicePaymentService.findById(invoicePaymentId);
+
+        checkPermission(payment);
+
         invoicePaymentService.deleteById(invoicePaymentId);
     }
 
-    public InvoicePayment saveInvoiceAttachment(long invoiceId, MultipartFile attachment) {
+    @Transactional
+    public InvoicePayment savePaymentAttachment(long invoiceId, @NonNull MultipartFile attachment) {
         final InvoicePayment payment = invoicePaymentService.findById(invoiceId);
+
+        checkPermission(payment);
 
         payment.setAttachment(fileUtils.compressFile(attachment, true));
         payment.setAttachment_name(attachment.getOriginalFilename());
@@ -70,8 +86,11 @@ public class InvoiceService {
         return invoicePaymentService.save(payment);
     }
 
-    public InvoicePayment removeAttachment(long invoiceId) {
+    @Transactional
+    public InvoicePayment removePaymentAttachment(long invoiceId) {
         final InvoicePayment payment = invoicePaymentService.findById(invoiceId);
+
+        checkPermission(payment);
 
         payment.setAttachment(null);
         payment.setAttachment_name(null);
@@ -79,8 +98,14 @@ public class InvoiceService {
         return invoicePaymentService.save(payment);
     }
 
+    @Transactional(readOnly = true)
     public byte[] getPaymentAttachment(long invoicePaymentId) {
         return invoicePaymentService.findById(invoicePaymentId)
                 .getAttachment();
+    }
+
+    private void checkPermission(final InvoicePayment payment) {
+        if (creditCardService.findById(payment.getCredit_card_id()).getUser_id() != utils.getAuthUser().getId())
+            throw new WithoutPermissionException();
     }
 }
